+++ packages/firstly/src/lib/svelte/server/index.ts
@@ -1,0 +1,1 @@
+export { handleCaching } from './handleCaching.js'